Fixed a few issues with saving data in db:
* Fixed an issue with wiring a  duplicate copy of the games in tables
after a game being edited.
* Fixed the parser to only respect ; after each clause, e.g. onDrop
shape1  hide shape2 play sound; shape2 goto page2;
* Added font size to the shape table.
* Added a column in script table to save the action order based on the
game document which states that for onClick "Executes the actions from
left to right." This is the last column (before _id) in the scripts game
and has the type integer. Also populated the hard-coded version of bunny
word with this column.